move the ND types also to the types folder
